Fix regex
Fixing regex for a case I detected where the json data had more fields in the map where "captionTracks" is situated.

Now "captionTracks" is not required to be the first nor the last field within this map. The curly brackets are being added later to make it a JSON-compatible data.